pyproject.toml
[project]
name = "kagami-bot-renderer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.2",
    "pika-stubs>=0.1.3",
    "pika>=1.3.2",
    "python-dotenv>=1.0.1",
    "selenium>=4.24.0",
]
